Module Battleship

module Author : sig ... end

CS 3110 Spring 2022 Final Project: Battleship

module Battleship_game : sig ... end

battleship_game.ml tracks the state of the game and contains functions for editing the state of the game.

module Command : sig ... end

Command enables parsing of player commands.

module Grid : sig ... end

grid.ml tracks the grid representation of the game, which includes every location on the grid and which locations have been bombed. Includes functions for placing ships and bombs on the grid.